+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes qsf-integration project will be documented in this file.
 ### Changed
 
 ### Deprecated
+- MonitoringBuilder.processingTime is deprecated and will be removed in the next release. Use the processingTimeFull or processingTimeUpdate instead
 
 ### Fixed